% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cognitive_decline.R
\docType{data}
\name{cognitive_decline}
\alias{cognitive_decline}
\title{Cognitive Decline data from Missouri residents (BRFSS).}
\format{
A data frame with 14,528 rows and 5 variables:
\describe{
\item{SEQNO}{key of the dataset}
\item{CIMEMLO1}{difficulties with thinking or memory}
\item{CIMEMLO1_IMP1}{1st imputation of CIMEMLO1}
\item{CIMEMLO1_IMP2}{2nd imputation of CIMEMLO1}
\item{CIMEMLO1_IMP3}{3rd imputation of CIMEMLO1}
}
}
\source{
\url{https://www.cdc.gov/brfss/index.html}
}
\usage{
cognitive_decline
}
\description{
A sub-dataset containing cognitive decline information from the
years 2023 and 2024.
}
\keyword{datasets}
