% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_demographics.R
\docType{data}
\name{demographics}
\alias{demographics}
\title{Demographic data from Missouri residents (BRFSS).}
\format{
A data frame with 14,528 rows and 14 variables:
\describe{
\item{SEQNO}{key of the dataset}
\item{MARITAL}{marital status}
\item{EDUCA}{education level}
\item{VETERAN3}{veteran status}
\item{EMPLOY1}{employment status}
\item{CHILDREN}{number of children in household}
\item{INCOME3}{income level}
\item{PREGNANT}{pregnancy status1}
\item{WEIGHT2}{reported weight in pounds}
\item{CRACE1}{child non-hispanic race including multiracial}
\item{CAGEG}{child age}
\item{AGE80}{age}
\item{GENHLTH}{general health}
\item{MENTHLTH}{mental health}
}
}
\source{
\url{https://www.cdc.gov/brfss/index.html}
}
\usage{
demographics
}
\description{
A sub-dataset containing demographics information from the
years 2023 and 2024.
}
\keyword{datasets}
